@@ -1,7 +1,7 @@
 import * as vscode from 'vscode';
 import * as path from 'path';
 import * as fs from 'fs';
-import AdmZip = require('adm-zip');
+import AdmZip from 'adm-zip';
 import { downloadFile } from '../utils/downloadHelper';
 
 /**
